+++ src/features/playback/store/playerStoreTypes.ts
@@ -189,7 +189,7 @@ export interface PlayerState {
     x: number;
     y: number;
     item: unknown;
-    type: 'song' | 'favorite-song' | 'album' | 'artist' | 'queue-item' | 'album-song' | 'playlist' | 'multi-album' | 'multi-artist' | 'multi-playlist' | null;
+    type: 'song' | 'favorite-song' | 'album' | 'artist' | 'queue-item' | 'album-song' | 'playlist' | 'multi-song' | 'multi-album' | 'multi-artist' | 'multi-playlist' | null;
     queueIndex?: number;
     playlistId?: string;
     playlistSongIndex?: number;
@@ -207,7 +207,7 @@ export interface PlayerState {
     x: number,
     y: number,
     item: unknown,
-    type: 'song' | 'favorite-song' | 'album' | 'artist' | 'queue-item' | 'album-song' | 'playlist' | 'multi-album' | 'multi-artist' | 'multi-playlist',
+    type: 'song' | 'favorite-song' | 'album' | 'artist' | 'queue-item' | 'album-song' | 'playlist' | 'multi-song' | 'multi-album' | 'multi-artist' | 'multi-playlist',
     queueIndex?: number,
     playlistId?: string,
     playlistSongIndex?: number,